requirements.txt
pytest
selene==2.0.0rc9
allure-pytest==2.13.5
python-dotenv
requests==2.32.3
jsonschema==4.22.0